Add JS code to randomly select one of the regions
This is triggered on `DOMContentLoaded`. In most cases, this should prevent people to see that the choice of region is actually taking place at runtime rather than being decided on the server.

Using client-side code instead of server-side logic may seem strange. The primary reason behind this choice is that it allows the compiled Blade templates to always stay the same, thus allowing caching. This seemed to be an important factor for the public home page of the site.