Make the link take the height of its content
The only child of the link is a floated <h1>, resulting in the link having a height of zero. In order to apply a hover effect on it, we need the link to have the height of its content. We thus float it in the same way as the heading it surrounds in order to make it retrieve a height.