Fix old tests that now need authentication
Now that the admin routes are protected by the `auth` middleware, we need to fix old tests that didn’t take that into account.

One of the test can simply be run by disabling all middleware.

The other test relies on route model bindings because the associated controller action expects a full Eloquent model, not just a slug. So, since we cannot simply disable middleware for this test, we create a team member and act as her while doing the request, so that the `auth` middleware is happy.