Fix capitalization of site title with a dedicated element
In order to fix observed erratic cross-browser behaviour regarding the `first-letter` pseudo-class/element, we now use a dedicated <span> element to isolate this letter and target it with a simple class selector.

While this solution is not as pure as the previous one, it is more consistent and should solve the issue once and for all.