Put site title on top of main menu on small viewports
Starting now, on small viewports, the default location of the site title and logo is above the main menu.

A special case is created for the home page, in order to keep the existing design there. This variant helps to make the logo and the title more prominent, which is perfect for that page.