Validate pre-launch partners who have no join date
So far, partners that were already in the currency network before the (re)launch but that did not have a `joined_on` date were not set as valid by the partner seeder. This commit fixes this and set the validation date so that it is the same as the (re)launch date.