Ensure a team retrieves the proper amount of members
This commit adds an assertion to an already existing test just to be extra sure that the method that is tested behaves as expected.

To be more precise, we ensure that, when a team gets the list of its members, the amount of retrieved members matches what is in the database.