Redirect to summary page when linking to a partner
Make the URLs that are directly pointing to a partner to redirect to this partner’s summary.

Using this, we can have simple URLs such as ‘/gestion/partneraires/foo’, and the step they point to will act as a ‘show’ page for the given partner.

Because of this change, we also need to update a feature test to ensure that it follows the HTTP redirection.