Work around a PHP bug (or hidden feature?)
It turns out that the null coalescing operator (`??`) does not try to call magic methods on objects.

This means that, if one tries to call a property that is not defined on an object, the operator will always consider that this property does not exist, *even when a call to `__get()` would return an actual non-null value*.

I don’t know if this surprising behaviour is intended or not but, for now, the simplest way to deal with it is to explicitly call `__toString()` on our object to get sure that the `??` operator inside `optional()` will work properly.