Replace ‘Under construction’ area by a list of links
These links are the same than the ones in the site header, the only differences being that they use a different, longer label.

These links are more or less there as placeholders. The main objective is to remove the ‘under construction’ area. The replacement content for this area may evolve in the future.