Allow Location to retrieve its partner’s public phones
This method returns the public phones of the location if there is at least one, otherwise it returns the public phones of the partner of the location.

If the partner itself has no public phone, an empty Collection is returned. As a result, the method is always expected to return a Collection.