Duplicate the middleware stack
In order to remove the use of cookies on public pages and still be able to use them (for sessions and CSRF protection) on admin pages, we will need to use different middlewares depending on the targeted URLs/pages.

A first step to enable this is to duplicate the stack of middleware and load one or the other depending on the URL that is being targeted.

This commit updates the route service provider in order to load a stack or the other.