Fix a test that got broken by the removal of cookies
There is a test in LogInToAdminAreaTest that verifies that members are redirected to the admin home page after login (if they did not try to go to a specific location inside the admin).

The problem is that the AuthenticatesUsers trait, which is provided by Laravel and is used in the login process, makes a call to the session to regenerate it.

It seems that removing the use of cookies on public pages had the unintended side effect of making this particular thing break, even though the login page is using cookies.

The solution applied here is to partially override the method that makes a call to the session, and to remove that call while keeping the rest of the method as is.

Manual testing has been done to ensure that the login process still works as expected.