Add custom view for HTTP 419 errors
This is a non-standard code that is used when a session token has expired (of if the corresponding cookie has been deleted).

It has been introduced in Laravel by this pull request:
https://github.com/laravel/framework/pull/18728

This error is usually very rare.

By providing a custom view, we ensure that people won’t get the default error page in English.