Prevent exception if one tries to GET the logout page
This should basically never happens under normal conditions. The only theoretical case where I could imagine this happening would be the following combo:

1. A logged in person clicks the ‘Log out’ button a great while after having loaded the page she is on (and the session token has expired). As a result, the person then gets a 419 error.

2. She tries to refresh the current URL, thus making a GET request on the logout URL, which only expects POST requests.

Very theoretical, but very misleading if it does happen. So here is a fix.

Thanks to this commit, instead of getting an exception for reaching an URL with an HTTP method that is not allowed, we will simply be redirected to the login form.