Rename `partners` relationship to `endorsedPartners`
We want to reassign this relationship name to retrieve the partners that have been *created* by a given team member.

By changing this name, we also make the relationship more clear, by explicitly saying that these partners are the ones that have been *endorsed* by a team member, but not necessarily created by her in the database.