Modify how formatting methods work
Starting from now, instead of having 2 different methods to format as string and 2 more to format as HTML, only one will exist to format as string (`toString()`) and only one to format as HTML (`toHtml()`).

The output of these two methods will be influenced by the value of the `$usePostalFormat` public property, which is false by default but can be switched to true by using the `asPostalMail()` modifier method.