Tweak migration to ‘fix’ SQLite-based tests
We currently use an in-memory SQLite database to execute PHPUnit tests, for performance reasons.

It turns out SQLite doesn't support multiple calls to `dropColumn` / `renameColumn` in a single modification. And one of the database migrations does exactly that.

So, in order to avoid breaking the tests, and since `down()` migration methods are, in fact, never used, we can tweak that migration by commenting the two `dropColumn()` statements.