Requalify `laravel/telescope` as a non-dev dependency
The application expects to find a service provider for Telescope, which is not present in production env because dev dependencies are not installed there. This causes the application to crash.

Since Telescope is built to properly handle cases where it’s installed in production environments (this is in fact the ‘recommended’ way), let’s do just that and requalify it as a non-dev dependency.