Signal cities having at least one currency exchange
In the list of cities, we now display a red-ish logo next to the name of cities having at least one currency exchange.

This logo uses ‘(C)’ as alt attribute in order to convey its information even when images are not available.

This commit updates the controller method, the Blade view, and some CSS. It also brings the SVG of the logo.