Add a new index page for partners
We recycle an old Blade view and replace its content by a few links and descriptions allowing to find partners in different ways.

The link to view partners with no location, as well as the note saying that some partners are not listed on the site at all, are both moved to this new index page and thus removed from the page listing the cities where there are partners.