Use generated JS file in the Blade view
Instead of getting data from the controller, we now inject the contents of the generated JS file into the view.

Since this generated file groups all data under a unique global variable (`window.generalMapData`), we also need to slightly update the main JS file that is building the map in order to properly retrieve the data.